Separation completed.
