+++ languages/wide-admin-menu-pt_BR.po
@@ -6,7 +6,7 @@ msgstr ""
 "PO-Revision-Date: 2024-09-18 01:35-0600\n"
 "Last-Translator: \n"
 "Language-Team: \n"
-"Language: es_MX\n"
+"Language: pt_BR\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
@@ -26,35 +26,35 @@ msgstr ""
 #: class.wide-admin-menu-settings-page.php:24
 #: class.wide-admin-menu-settings-page.php:74
 msgid "Wide Admin Menu"
-msgstr "Menú Admin Ancho"
+msgstr "Menu de administração amplo do WP"
 
 #: class.wide-admin-menu-settings-page.php:44
 msgid "Settings Saved"
-msgstr "Ajustes guardados"
+msgstr "Configurações Salvas"
 
 #: class.wide-admin-menu-settings-page.php:58
 msgid "Save Settings"
-msgstr "Guardar Ajustes"
+msgstr "Salvar Configurações"
 
 #: class.wide-admin-menu-settings-page.php:75
 msgid "Settings for the Wide Admin Menu plugin."
-msgstr "Configuración del plugin Wide Admin Menu."
+msgstr "Configurações para o plug-in Wide Admin Menu."
 
 #: class.wide-admin-menu-settings-page.php:85
 msgid "General"
-msgstr "General"
+msgstr "Geral"
 
 #: class.wide-admin-menu-settings-page.php:91
 msgid "Width"
-msgstr "Ancho"
+msgstr "Largura"
 
 #: class.wide-admin-menu-settings-page.php:115
 msgid "General settings for the Wide Admin Menu plugin."
-msgstr "Configuración general del plugin Wide Admin Menu."
+msgstr "Configurações gerais para o plug-in Wide Admin Menu."
 
 #: class.wide-admin-menu-settings-page.php:124
 msgid "The width for the Admin menu."
-msgstr "La anchura del menú Admin."
+msgstr "A largura do menu Admin."
 
 #. Plugin URI of the plugin/theme
 msgid "https://rhaseven7h.com/wordpress-wide-admin-menu-plugin/"
@@ -62,7 +62,7 @@ msgstr "https://rhaseven7h.com/wordpress-wide-admin-menu-plugin/"
 
 #. Description of the plugin/theme
 msgid "A simple plugin for WordPress to widen the Admin menu."
-msgstr "Un simple plugin para WordPress para ampliar el menú de administración."
+msgstr "Um plug-in simples para o WordPress para ampliar o menu de administração."
 
 #. Author of the plugin/theme
 msgid "Gabriel Medina"
